pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yolov9001"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "pyyaml",
    "albumentations>=1.3.1",
    "torchvision",
    "torch",
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "utils*", "core*"]